requirements.txt
fastapi
uvicorn
pydantic>=2.7.0
pydantic-settings
sqlmodel
aiomysql
pymysql
langchain
langchain-community
langchain-openai
langchain-chroma
python-dotenv
pypdf
docx2txt
sentence-transformers
python-jose[cryptography]
alembic
sentry-sdk
